+++ packages/pigeon/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 25.0.0
 
+* **Breaking Change** Removes `oneLanguage` field from `PigeonOptions`.
 * Separates internal options classes from user facing options.
 * Adds `mergeDefinitionFileOptions` parameter to `runWithOptions` method.
 * Relocates `injectOverflowTypes` to non-public facing method.
