+++ packages/pigeon/test/cpp_generator_test.dart
@@ -5,7 +5,7 @@
 import 'package:pigeon/src/ast.dart';
 import 'package:pigeon/src/cpp/cpp_generator.dart';
 import 'package:pigeon/src/generator_tools.dart';
-import 'package:pigeon/src/pigeon_lib_internal.dart' show Error;
+import 'package:pigeon/src/pigeon_lib.dart' show Error;
 import 'package:test/test.dart';
 
 const String DEFAULT_PACKAGE_NAME = 'test_package';
